WIP: add Saml2KeystoneUnscoped for K2K federation
Implement the auth plugin noted for K2K federation. The existing
Saml2ScopedToken plugin works for getting a scoped token.

This is marked WIP for 3 reasons:
1) Unit tests (I'm working this)
2) I wonder if this should merge in function defined in
https://review.openstack.org/#/c/159022/
so that the API user doesn't have to get the SAML at all.
Instead this plugin could be enhanced to just accept a scoped token
for the IDP, could perform the samlization as well, and return the
unscoped token for the SP.
3) I'm a bit unclear about lxml usage as noted by TODOs in the code

Change-Id: I6fc99882a7733f138c305cf3b48bd34be6ed4e2c